im losing my fucking mind

+++ dzemCiastkowy/Source/Game/NPC/FriendlyNpcInstance.cs
@@ -8,5 +8,5 @@ public class FriendlyNpcInstance(NpcData data)
 
     public int maxHealth = data.friendMaxHealth;
     public int health = data.friendHealth;
-    public int actionPoints = 0;
+    public float actionPoints = 0;
 }
